services/python/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.1
django-allauth==64.2.1
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
kombu==5.4.1
oauthlib==3.2.2
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13


#in order to run the celery(i,e in order to make realtime trackig of auction status )you should install redis and run this two commands
# 1. celery -A bidding_system worker --loglevel=info
# 2. celery -A bidding_system beat --loglevel=info